Add communication with API. Fix Favorite.

+++ products_app/lib/pages/product_list.dart
@@ -1,10 +1,25 @@
 import 'package:flutter/material.dart';
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:scoped_model/scoped_model.dart';
 
 import './product_edit.dart';
 import '../scoped-models/main.dart';
 
-class ProductListPage extends StatelessWidget {
+class ProductListPage extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() {
+    return _ProductListPageState();
+  }
+}
+
+class _ProductListPageState extends State<ProductListPage> {
+  @override
+  void initState() {
+    MainModel model = ScopedModel.of(context);
+    model.fetchProducts();
+    super.initState();
+  }
+
   Widget _buildEditButton(BuildContext context, int index, MainModel model) {
     return IconButton(
       icon: Icon(Icons.edit),
@@ -26,15 +41,28 @@ class ProductListPage extends StatelessWidget {
     return ScopedModelDescendant<MainModel>(
       builder: (BuildContext context, Widget child, MainModel model) {
         final products = model.products;
+        if (products == null) {
+          return Center(
+            child: CircularProgressIndicator(),
+          );
+        }
+        if (products.length == 0) {
+          return Expanded(
+            child: Container(
+              margin: EdgeInsets.only(top: 10.0),
+              child: Text('No produts'),
+            ),
+          );
+        }
         return ListView.builder(
           itemBuilder: (BuildContext context, int index) {
             return Dismissible(
               key: Key(products[index].title),
               direction: DismissDirection.endToStart,
-              onDismissed: (DismissDirection direction) {
+              onDismissed: (DismissDirection direction) async {
                 if (direction == DismissDirection.endToStart) {
                   model.selectProduct(index);
-                  model.deleteProduct();
+                  await model.deleteProduct();
                   model.selectProduct(null);
                 }
               },
@@ -45,7 +73,8 @@ class ProductListPage extends StatelessWidget {
                 children: <Widget>[
                   ListTile(
                     leading: CircleAvatar(
-                      backgroundImage: AssetImage(products[index].image),
+                      backgroundImage:
+                          CachedNetworkImageProvider(products[index].image),
                     ),
                     title: Text(products[index].title),
                     subtitle: Text('\$${products[index].price.toString()}'),
